+++ packages/blade/src/components/Icons/PettyCashBudgetIcon/PettyCashBudgetIcon.web.test.tsx
@@ -1,0 +1,11 @@
+import PettyCashBudgetIcon from './';
+import renderWithTheme from '~utils/testing/renderWithTheme.web';
+
+describe('<PettyCashBudgetIcon />', () => {
+  it('should render PettyCashBudgetIcon', () => {
+    const { container } = renderWithTheme(
+      <PettyCashBudgetIcon color="feedback.icon.neutral.intense" size="large" />,
+    );
+    expect(container).toMatchSnapshot();
+  });
+});